Move utility js files into their own folder
