+++ frontend/src/components/Dashboard.jsx
@@ -1,7 +1,7 @@
 import React, { useState, useEffect, useCallback, useRef   } from 'react';
 import { Container, Row, Col, Card, CardTitle, CardBody } from 'reactstrap';
 import { DASHBOARD_API_URL, JOB_POSTING_API_URL, JOB_OPPORTUNITY_API_URL } from "../constants";
-import { useApiRequest } from '../useApiRequest'; // Import the hook
+import { useApiRequest } from '../utils/useApiRequest'; // Import the hook
 import { formatDisplayDate } from '../constants'; 
 import DataTableBase from './DataTableBase';
