+++ frontend/src/components/JobSiteEdit.jsx
@@ -8,7 +8,7 @@ import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faTrash, faFloppyDisk } from '@fortawesome/free-solid-svg-icons'
 
 import Editor from "./Editor"
-import { useApiRequest } from "../useApiRequest";
+import { useApiRequest } from "../utils/useApiRequest";
 
 const JobSiteEdit = () => {
