+++ frontend/src/components/JobSiteList.js
@@ -1,6 +1,6 @@
 import React, { useEffect, useState, useCallback, useRef } from 'react';
 import { Link, useNavigate } from 'react-router-dom';
-import { useApiRequest } from '../useApiRequest';
+import { useApiRequest } from '../utils/useApiRequest';
 import { JOB_SITE_API_URL, formatDisplayDate  } from "../constants";
 
 import DataTableBase from './DataTableBase';
