+++ frontend/src/components/OpportunityDetails.jsx
@@ -6,7 +6,7 @@ import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faTrash, faFloppyDisk } from "@fortawesome/free-solid-svg-icons";
 import Editor from "./Editor";
 import Comments from "./Comments";
-import { useApiRequest } from "../useApiRequest";
+import { useApiRequest } from "../utils/useApiRequest";
 
 const OpportunityDetails = () => {
   const [state, setState] = useState({
